User: Make the `user_login` field `readonly` instead of `disabled`.
On the profile edit screen in ``wp-admin, the username was previously displayed using an input field with the `disabled` attribute. This field was skipped by screen readers, which leads to a loss of information for visually impaired users. This changeset switches it to a `readonly` attribute, which is exactly the purpose of this field.

Props roytanck, audrasjb, rajinsharwar, joedolson.
Fixes #60774.




git-svn-id: https://develop.svn.wordpress.org/trunk@58280 602fd350-edb4-49c9-b593-d223f7449a82