Editor: Add labels for template name display in site editor.
Add the label `template_name` to post types and taxonomies for use by the site editor to display on the templates/add new template screens to allow extenders to customize the display.

For post types the template name applies to templates for a singular pages with the default value "Single item: [singular name]".

For taxonomies the template name applies to templates for term archives with the default value "[singular name] Archives".

Props aljullu, audrasjb, ntsekouras, ellatrix, oglekler, rajinsharwar.
Fixes #60881.



git-svn-id: https://develop.svn.wordpress.org/trunk@58377 602fd350-edb4-49c9-b593-d223f7449a82