Twenty Twenty-Three: Remove untranslatable text strings in pagination links.
This changeset removes text strings from the query pagination block that are not translatable without adding them to patterns.
This means that the texts now use the default translation for the block from WordPress, not the theme. The archive, search and home templates now match the text in the query pagination in the index template.

Props poena.
Fixes #60298.




git-svn-id: https://develop.svn.wordpress.org/trunk@58459 602fd350-edb4-49c9-b593-d223f7449a82