Twenty Fourteen: Resolves post content links not being underlined.
In the page list block the text decoration was not the same on the front as the editor. This adds underline to links when in the post editor classic or block. This does not resolve the font size issue that needs to be fixed in block editor reset styles.

Props viralsampat, sabernhardt.
Fixes #61441.


git-svn-id: https://develop.svn.wordpress.org/trunk@58463 602fd350-edb4-49c9-b593-d223f7449a82