Twenty Fourteen: Updates styles for Table block and figure captions.
There were multiple discrepancies between the editor and front end with [58399] matched incorrect border color due to specificity. This resolves that and also includes issues not caught in [60293]. The full list of what this does is in ticket but a summary is reduces specificity for table cell border, removes border color rules, replaces selectors and adds wrapper so alignment changes within the iframe.

Props sabernhardt.
Fixes #61563.


git-svn-id: https://develop.svn.wordpress.org/trunk@58686 602fd350-edb4-49c9-b593-d223f7449a82